Implement Ramp mode
Ramp mode allows the driver to crawl the left or the right side of the
robot.  Ramp mode is enabled by depressing button 4 on the stick to
crawl to the left (right motors engaged) or button 5 on the stick to
crawl to the right (left motors engaged).  The buttons match the side
the drive would push the stick to get the same effect.  X axis
effectively mirrors the Y axis when Ramp mode is active.

For example, if the left ramp sensor is green, then use button 4
to crawl the right side of the robot forward until the right
ramp sensor is green.