Implement Sensors object to bridge reads with LED updates
The WPILib scheduler will cancel any pending commands that require
the same resources as the command being added.  This was causing the
UpdateEventLightsCommand to cancel pending light updates.

This commit moves the sensor state into its own object and uses it
as a collection point.  The UpdateEventLightsCommand is now "slaved"
to the Sensors object state.